Merges and resolutions
